Move verify domino support to own module
