+++ domino/aisystems/tracing/_util.py
@@ -1,0 +1,5 @@
+import os
+
+def get_is_production() -> bool:
+    return os.environ.get("DOMINO_AI_SYSTEM_IS_PROD", "false").lower() == "true"
+
