+++ tests/aisystems/test_ai_systems_utils.py
@@ -1,51 +1,5 @@
-from unittest.mock import patch
-import os
-import pytest
-
-from domino.aisystems._constants import MIN_MLFLOW_VERSION, MIN_DOMINO_VERSION
-from domino.aisystems._util import _get_version_endpoint,  build_eval_result_tag
-from domino.exceptions import UnsupportedOperationException
-from ..conftest import TEST_AI_SYSTEMS_ENV_VARS
-
-def test_get_version_endpoint():
-        with patch.dict(os.environ, TEST_AI_SYSTEMS_ENV_VARS | {"DOMINO_API_HOST": "http://localhost:1111/"}, clear=True):
-                assert _get_version_endpoint() == "http://localhost:1111/version"
+from domino.aisystems._util import build_eval_result_tag
 
 def test_build_eval_result_tags():
         assert build_eval_result_tag('my_metric', '1') ==  'domino.prog.metric.my_metric', 'numbers should be metrics'
         assert build_eval_result_tag('my_label', 'cat') ==  'domino.prog.label.my_label', 'strings should be labels'
-
-def test_verify_domino_support_domino_and_mlflow_correct_version(verify_domino_support_fixture):
-        from domino.aisystems._util import verify_domino_support
-        verify_domino_support_fixture['mock_get_domino_version'].return_value = MIN_DOMINO_VERSION
-        verify_domino_support_fixture['mock_get_mlflow_version'].return_value = MIN_MLFLOW_VERSION
-
-        # Should not raise
-        verify_domino_support()
-
-def test_verify_domino_support_domino_wrong_version(verify_domino_support_fixture):
-        from domino.aisystems._util import verify_domino_support
-        verify_domino_support_fixture['mock_get_domino_version'].return_value = "6.1.2"
-
-        with pytest.raises(UnsupportedOperationException) as exn:
-                verify_domino_support()
-
-        assert str(exn.value) == "This version of Domino doesn’t support the aisystems package."
-
-def test_verify_domino_support_mlflow_wrong_version(verify_domino_support_fixture):
-        from domino.aisystems._util import verify_domino_support
-        verify_domino_support_fixture['mock_get_domino_version'].return_value = MIN_DOMINO_VERSION
-        verify_domino_support_fixture['mock_get_mlflow_version'].return_value = '3.1.0'
-
-        with pytest.raises(UnsupportedOperationException) as exn:
-                verify_domino_support()
-
-        assert str(exn.value) == f"This code requires you to install mlflow>={MIN_MLFLOW_VERSION}"
-
-@pytest.fixture
-def verify_domino_support_fixture():
-        with patch.dict(os.environ, TEST_AI_SYSTEMS_ENV_VARS | {"DOMINO_API_HOST": "http://localhost:1111/"}, clear=True), \
-                patch('domino.aisystems._util._get_domino_version') as mock_get_domino_version, \
-                patch('domino.aisystems._util._get_mlflow_version') as mock_get_mlflow_version:
-                yield { 'mock_get_domino_version': mock_get_domino_version, 'mock_get_mlflow_version': mock_get_mlflow_version }
-
